Wait for lifecycle update to eliminate flakiness in VisualPropertiesSynchronization.RemoteToLocalTransition
This test was flaky on Fuchsia and ChromeOS Ash. The theory is that we
need to wait longer for the VisualProperties update to come in. Waiting
for a lifecycle update will hopefully ensure that update has happened.

Bug: 361299696
Change-Id: I2479f5f8b7daab68f05e3c23ec0b477e6d63640d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5857258
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Reviewed-by: Brandon Maslen <brandm@microsoft.com>
Auto-Submit: Erik Anderson <Erik.Anderson@microsoft.com>
Commit-Queue: Dave Tapuska <dtapuska@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1355281}