[EducationalTip] Polish DefaultBrowserPromoCoordinator.
In this CL, we enhance the DefaultBrowserPromoCoordinator class by introducing a destroy function and additional comments for clarity. Furthermore, we address a scenario where the card type is not supported. In such cases, the ModuleDelegate#onDataFetchFailed function will be invoked to remove the educational tip module effectively.

Bug: 355015904
Change-Id: I32f504bda8174bc1707c2978e1a3282ef709f9e8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5854275
Reviewed-by: Xi Han <hanxi@chromium.org>
Commit-Queue: Xinyi Ji <xinyiji@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1355322}