Change Observer: Fix crash when deleting large folder
The Inotify FilePathWatcher can report an event out of its scope if the
watched directory is deleted. When the watched directory is deleted, it
clears out the recursive watches, but its possible that it does that
right before reporting an event for a descendant.

This fixes it by calculating the reported path before clearing out the
recursive watches.

Fixed: 365839167
Change-Id: Ib5d487e236bc6ab80840c94896c5c980e183a873
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5852175
Reviewed-by: Christine Hollingsworth <christinesm@chromium.org>
Commit-Queue: Nathan Memmott <memmott@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1355330}