Replace BrowserList with WindowControllerList
Replace the straightforward uses of BrowserList in
//chrome/browser/extensions with WindowControllerList.

This is work in support of eventually removing the Browser dependency
from the extensions system.

Replace users of chrome::GetBrowserCount() in
//chrome/browser/extensions
with a new function GetWindowControllerCountInProfile that is specific
to the extensions system and doesn't depend on BrowserList.

Bug: b/361838438
Change-Id: I0cae38d57faf31d64b48a292ba0937f70d5180ba
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5847308
Reviewed-by: Finnur Thorarinsson <finnur@chromium.org>
Commit-Queue: Brett Wilson <brettw@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1355372}