webauthn: warn about attestation rather than prompting about it.
Traditionally, Chrome has shown a prompt when a site requests
attestation. This was done to try and ensure that WebAuthn didn't end up
where only a handful of major security keys worked on the public
internet because of the limited intersection of varying attestation
requirements by sites.

But we're past that point now. Passkey providers are widely supported
and don't provide any attestation. This has ensured that sites aren't
blocking registrations based on attestation.

On the other hand, the attestation prompts are bothering enterprise
cases which attestation was designed for.

Additionally, registrations over hybrid currently always strip
attestation and remove the AAGUID. But this contrasts with registering
with a passkey provider locally, where the AAGUID is often used to show
the user the passkey provider in settings screens. This means that
registrations over hybrid work differently from local registrations,
which is unfortunate and so that is also addressed in this change.

Windows webauthn.dll v1 didn't support showing an attestation warning,
but that version is very old now and so we don't keep all the
attestation UI infrastructure around just for that case.

Fixed: 365818841
Change-Id: I7fc71ca922fabda81562072bc6d1b2bd730bc6a0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5842600
Commit-Queue: Adam Langley <agl@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Mohamed Amir Yosef <mamir@chromium.org>
Reviewed-by: Martin Kreichgauer <martinkr@google.com>
Cr-Commit-Position: refs/heads/main@{#1355382}