[iOS] Add more information in CHECK in HistorySyncMediator
AuthenticationService::GetPrimaryIdentity() can return nil according
to crbug.com/366198713.

This doesn't possible since the mediator listens for IdentityManager
to know when the primary account disappear. When it happens,
the mediator sends the signal to the coordinator to dismiss itself.

Two possible issues seems possible:
1. There is a mismatch between IdentityManager and
   AuthenticaitonService.
2. The history sync opt-in dialog is not removed on sign out.

This patch update CHECK to add in the log if there is primary
account in AuthenticationService and IdentityManager to understand
if 1- is possible.

When the primary identity is removed to device, 1. could be explained
by a race condition. The id<SystemIdentity> for the primary account
is removed from ChromeAccountManagerService, and IdentityManager
didn't update itself yet.
Therefore, the mediator didn't receive the notification yet, and
AuthenticationService cannot find id<SystemIdentity>.

This seems very unlikely but it would be interesting to prove if
this is the cause of this CHECK failure.

Bug: 366198713
Change-Id: I1543c937f904ef8a0251842e4b9b650b0af37084
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5860423
Auto-Submit: Jérôme Lebel <jlebel@chromium.org>
Commit-Queue: Jérôme Lebel <jlebel@chromium.org>
Reviewed-by: Marc Treib <treib@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1355386}