Reduce fragility in SidePanelCoordinator (part 6 / N)
This CL is a refactor with no intended behavior change. This CL changes
SidePanelContentSwappingContainer to use a cancellable callback. This
simplifies the interface for SidePanelContentSwappingContainer and
allows removal of production logic intended to deal with non-cancelled
but outdated callbacks.

Change-Id: I656ca0fc30379ee4924cd08de5d6f7d513ac8d66
Bug: 363743081
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5841167
Commit-Queue: Erik Chen <erikchen@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Caroline Rising <corising@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1355422}