Use Parsed Forms For Form Event Logging
Instead of logging to each form type of a form, we log for each of the parsed form types on a page. This is to fix the submission metric regression where we do not log submit metrics across all forms on the webpage.

Bug: 359934323
Change-Id: I9477e180cc0a867cb6348fd4aef5e5567629bb37
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5854942
Commit-Queue: Alexander Tekle <alexandertekle@google.com>
Reviewed-by: Dominic Battré <battre@chromium.org>
Reviewed-by: Siyu An <siyua@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1355432}