Add ChromeDriver support to bisect-builds.py.
For automated tests that depend on WebDriver/ChromeDriver like Selenium
and 1P tests, having the correct version of ChromeDriver is necessary.
Therefore, we want the script to be able to download ChromeDriver during
bisection.

This CL adds --chromedriver support for desktop builds, and the path to
ChromeDriver can be passed to `--command` via the `"%d"` placeholder.

Bug: 352344089
Change-Id: Ide18ee115adcb0151770d2d0010d60415789b960
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5861048
Commit-Queue: Kuan Huang <kuanhuang@chromium.org>
Reviewed-by: Sven Zheng <svenzheng@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1355470}