webauthn: disable iCloud Keychain passkeys in PWAs.
PWAs, in order to appear as 1st-class apps on macOS, render their UI in
a separate process. The iCloud Keychain APIs require an NSWindow* in
order to position their dialogs over the correct application window.
But, for PWAs, the correct NSWindow* is in a different process and so
cannot be used. We would likely have to perform the iCloud Keychain
calls in the remote process for things to work.

That involves a non-trivial amount of work to define the Mojo interface,
marshal and unmarshal all the parameters, etc. We are not able to do
that immediately, but iCloud Keychain is currently comically broken.
(The dialog appears under the PWA window and will "run away" if you try
to uncover it.)

Thus this change disables iCloud Keychain in PWAs for now.

Change-Id: I86d502c5c1f510752f1aa91e8acacf855da91df3
Bug: 364926914
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5861477
Reviewed-by: Martin Kreichgauer <martinkr@google.com>
Commit-Queue: Adam Langley <agl@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1355491}