focus-mode: Fix clicking Get Premium button after uninstalled
This CL fix a bug that after the user uninstalled the web app but
without reboot for the device, clicking on the Get premium button, no
chrome browser was launched.
After the changes, we will launch the browser if we launched the web app
failed.

Bug: b/366558524
Test: Manual
Change-Id: I59ea87494a33e18742acf99cc966d47e9b6cf0bf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5862013
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>
Reviewed-by: Richard Chui <richui@chromium.org>
Commit-Queue: Hongyu Long <hongyulong@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1355518}