Modify Me2MeAuthenticatorFactory to check the current host_owner_
This CL is needed to support server-provided host owners and updates.
Prior to this change, the host_owner email in the config was provided
to the Me2MeAuthenticatorFactory which stored a copy. Even if the
service provided a new owner (or in the future, a list of owners
or domains), the authenticator factory was unaware of those changes.

In this CL, I'm now passing in an authz callback that the
authenticator can call to see if the user email in question is an
owner and if its domain is included in the client domain policy.

Bug: 346825839
Change-Id: Id5601d73959314e4f9374963992573224e9d7d1d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5861474
Reviewed-by: Gary Kacmarcik <garykac@chromium.org>
Commit-Queue: Joe Downing <joedow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1355526}