Update FtlEchoMessageListener to use a callback to check access
Similar to the previous CL, this one updates the class to store
a callback which is used when an echo message is received. One
difference between the existing impl and the change in this CL is
that messages will be rejected if a message was received from
an account whose domain was not in the allowlist. I think this
is a reasonable change though as presumably we wouldn't want
to respond if the client is not able to connect.

Bug: 346825839
Change-Id: I2bcb945cfd420544961bc150e557eb8441be75ca
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5862136
Reviewed-by: Gary Kacmarcik <garykac@chromium.org>
Commit-Queue: Joe Downing <joedow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1355528}