[CrOS Network] Request update if receive network state property update
This is a similar fix as crrev/c/5304120 for b/316312211, but for the
network state.
The initial GetProperties value may get stale if Chrome receives Shill
update immediately after the initial GetProperties sends to Shill, but
before Chrome finishes processing the response and set receive_update().
So, we need to request another round of GetProperties in this case to
get the latest properties value.

Bug: b/331845752
Change-Id: Icf127d13066c4526a165011772451480a60850c0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5824500
Commit-Queue: Jason Zhang <jiajunz@google.com>
Reviewed-by: Chad Duffin <chadduffin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1355548}