Remove feature AutofillDontPreserveAutofillState
Autofill has two versions of "AutofillState":

(1) The state stored in the renderer, maintained by both Autofil and
    blink and directly affected by autofill/user/JS modifications.

(2) The state stored in the browser, maintained by autofill only and
    affected by the signals we receive from the renderer and from our
    UIs.

Currently, we sync them the following way:

- On autofill, we set both of the states to Autofilled.

- On user edits of autofilled fields, we set both of the states to
  NotAutofilled.

- On JS edits of autofilled fields, we set only (1) to NotAutofilled
  but keep (2) as Autofilled.

- At form submission time ONLY, we set (2) to (1), whatever (1) is.

This means that, during form parsing at least, (1) and (2) can be out
of sync. This was considered as a problem and this feature was
introduced as a possible fix: Always set (1) to (2).

This is however not optimal. (1) is very hazardous and can change for
so many reasons, and the value of this state is crucial for our metrics
and flows, which means we can't really rely on what JS runs in the
renderer in order to log metrics and define flows.

Also note that the feature rolled out twice and failed twice because of
very strange metric behavior that are justified by the paragraph above.

The CL therefore removes the feature. Another feature currently rolling
out is AutofillFixCachingOnJavaScriptChanges, and after launching, we
might wanna reconsider whether maybe we want the different direction:
Never set (2) to (1) (At neither parsing nor submission time). This
would, for example, fix scenarios where:

- A user autofills a fields
- JS edits it
- The user performs no further edits and submits
- Autofill marks this field (in metrics) as ManuallyFilled or Corrected

And ultimately this would make it easier to reason about the
AutofillState at submission time.

Bug: 40871691
Change-Id: I8b8fb9e10cc690432736cc8826116dc7e1ca1484
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5860133
Reviewed-by: Christoph Schwering <schwering@google.com>
Commit-Queue: Christoph Schwering <schwering@google.com>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Auto-Submit: Jihad Hanna <jihadghanna@google.com>
Cr-Commit-Position: refs/heads/main@{#1355601}