Update ScreenAI's handling of loaded files when service disconnects.
If the ScreenAI service gets disconnected while loading the model files,
service initialization functions are not called anymore and the handles
of the opened model files are passed to a thread that supports blocking
to close them.

`PerformOCR_AfterServiceRevive` test that was flaky because of this
issue is re-enabled.

Cq-Include-Trybots: luci.chromium.try:linux-blink-web-tests-force-accessibility-rel

AX-Relnotes: n/a
Bug: 362478055
Change-Id: I03b64f680efba8df64b33e1b1bb1e19c8d10a116
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5853661
Reviewed-by: David Tseng <dtseng@chromium.org>
Commit-Queue: Ramin Halavati <rhalavati@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1355718}