[discard] Update MemorySaver tests to run with new discard impl
This CL updates MemorySaver interactive tests to run with both
the current and new discard impl.

The WaitForHide() step in DiscardAndReloadTab() was removed as
this only applies to the current impl and is unnecessary as
the WebContents is destroyed synchronously with the discard
request itself.

Bug: 347770670, 40893068
Change-Id: I14da7be58b1c2a673b24690cd8726653edcb8640
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5850911
Reviewed-by: Dana Fried <dfried@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Eshwar Stalin <estalin@chromium.org>
Commit-Queue: Thomas Lukaszewicz <tluk@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1355755}