[UNOp2][Android] Add UI for add account flow in tests
Previously, the add account flow in test was replaced by a stub activity
AddAccountActivityStub that finishes automatically after being created,
and sends a result that is set before the activity is shown.

This CL adds a layout with ok and cancel buttons to this activity, so
tests can interact directly with the activity's UI instead of using the
flow described above. It also updates existing usage of the flow
accordingly.

The change:
* better simulates the real add account flow which brings an activity to
the front;
* allows more cases to be tested (eg. previous activity killed during the
add account flow);
* improves consistency with iOS test infrastructure.

Fixed: 350466139
Change-Id: I5e2df90254cf26c0c440c7b75fa7364817ac831f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5831366
Commit-Queue: Menghan Yang <myuu@google.com>
Reviewed-by: Boris Sazonov <bsazonov@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1355759}