Account menu error item: Fix background in dark mode
The error item in the account menu should have the same background color
as everything else in the menu. Before this CL, its background was set
to kGroupedSecondaryBackgroundColor, which evaluated to a different
color only in dark mode, likely just accidentally.
This CL removes the explicit color, making the item match the rest of
the menu.

Bug: 365940219
Change-Id: Ia393c1c5986b73a7038ef4a2b04e9320d6412a1e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5864449
Commit-Queue: Marc Treib <treib@chromium.org>
Reviewed-by: Jood Hajeer <jood@google.com>
Cr-Commit-Position: refs/heads/main@{#1355768}