Fenced frames: Update shared storage get error message when feature FencedFramesLocalUnpartitionedDataAccess is disabled.
The current error message when fenced frame feature is on but
FencedFramesLocalUnpartitionedDataAccess feature is off, and shared
storage get is called in a fenced frame:
"Cannot call get() outside of a fenced frame."

This is confusing because the call did happen in a fenced frame.

Change it to explain that the failure is due to the local unpartitioned
data access feature is off.

Bug: 324440086
Change-Id: Ic8e29c994f6abba851f031a2263cbec171eb957c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5854088
Reviewed-by: Liam Brady <lbrady@google.com>
Commit-Queue: Xiaochen Zhou <xiaochenzh@chromium.org>
Reviewed-by: Yao Xiao <yaoxia@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1355785}