Fenced frames: Interrupt in-progress "Save Link As..." download request
This is a follow-up of crrev.com/c/5564165.

In this CL, the nonce for subresource is attached to the IsolationInfo
of the download request from "Save Link As...".

Other download options "Save Video As..", "Save Image As.." and
"Save Audio As..." already have the nonce attached when formulating
the download requests.

This allows the interruption of in-progress "Save Link As..."
download request when fenced frame disables untrusted network in
middle of the download.

Add a test that simulates a slow download and the fenced frame cuts
off untrusted network access before the download finishes. The
download is interrupted.

Moved tests from download_browsertest.cc to download_interactive_uitest.cc as browser_tests are run in parallel
and if a window loses focus the context menu will disappear.

Bug: 1515599
Change-Id: Ia0f053deb7670adf176e446fe85ca7003c8f95b4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5667085
Reviewed-by: Erik Chen <erikchen@chromium.org>
Commit-Queue: Xiaochen Zhou <xiaochenzh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1355790}