[ios] Improve SessionRestorationServiceImpl
Before this CL, if an unrealized WebState was detached from a Browser,
and the session saved before it could be inserted into another Browser
then the code would crash as it would be unable to find the original
Browser that contained the data.

Change SessionRestorationServiceImpl to keep the OrphanInfo for those
detached WebStates until they are either closed, or inserted into a
Browser.

There was no way to exercise this crash before as the production code
would always move a WebState between Browser without returning to the
main loop, but it was theoretically possible. The hole is now fixed.

Add unit tests that this fixes the crash during both the session save
and the LoadWebStateStorage(...) method.

Bug: none
Change-Id: Iaa9ec3709d2e69d92e35b429224cc0614ea75688
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5860161
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1355821}