[Android][3PPWM] Guard 3P Autofill in tabs with pref check
Without this CL, enabling the AutofillVirtualViewStructureAndroid
feature suppresses 3P AutofillServices' attempts to use the
accessibility API for Android Autofill. This API is the legacy
approach to Autofilling Chrome. Instead of deprecating that legacy
API suddenly, control it with a new feature, added in [1].

With this CL, the autofill.using_virtual_view_structure pref is also
checked to assess readiness to use VirtualViewStructures.
The pref is set as soon as the first autofill client is initialized [2].
The value of the pref is retrieved from a helper [3] that checks vital
preconditions like:
* The user opted into the autofill.using_virtual_view_structure setting,
* An AutofillService must be configured in android settings,
* The AutofillService can't be Autofill with Google,
* Android Autofill must be supported in general, and
* No policy prevents use of 3P AutofillServices.

(See https://developer.android.com/identity/autofill for more
information on AutofillServices and their interaction with apps.)

[1] https://chromium-review.googlesource.com/c/chromium/src/+/5844874/15
[2] https://source.chromium.org/chromium/chromium/src/+/main:chrome/browser/ui/autofill/autofill_client_provider.cc;l=93;drc=121aa1ab60d6beaa7253ad62dba962cc2c4a35a0
[3] https://crsrc.org/c/chrome/browser/autofill/android/java/src/org/chromium/chrome/browser/autofill/AutofillClientProviderUtils.java;l=49;drc=121aa1ab60d6beaa7253ad62dba962cc2c4a35a0

go/3ppwm-dd

Bug: b/326231439
Change-Id: I459c1ace6283a4e699babd0aa8e32e95603d5eaa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5860242
Reviewed-by: Calder Kitagawa <ckitagawa@chromium.org>
Reviewed-by: Dominic Battré <battre@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Friedrich Hauser <friedrichh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1355824}