[A11y] Deflake tests that use display: inline-block
Element count was used in attempt to reduce the number of
display:inline-block elements that are included in the tree.
The optimization does not affect behavior, but led to flakiness
because the page load can pause in the middle of the parent.

The optimization is likely unnecessary complexity -- web content
does not typically use display:inline-block.

Bug: 365027678
Change-Id: Idbfacb42ab9aa7332dc916c4e7dfcc4764aea259
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5859350
Auto-Submit: Aaron Leventhal <aleventhal@chromium.org>
Reviewed-by: Mark Schillaci <mschillaci@google.com>
Commit-Queue: Aaron Leventhal <aleventhal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1355832}